REFACTOR: Complete export and client routes migration
- Implemented comprehensive export blueprint with CSV export functionality for:
  - Tasks export with full metadata (assignees, projects, clients)
  - Projects export with progress tracking and statistics
  - Clients export with project counts and contact details
- Enhanced clients blueprint with complete CRUD operations:
  - edit_client, delete_client, mark_client_inactive routes
  - Comprehensive cascade deletion handling for client data
  - Project status management tied to client activation
- Reduced app.py routes from 69 to 63 (6 more routes migrated)
- Total progress: 37 routes migrated (100 → 63 remaining)
