+++ app.py
@@ -776,297 +776,9 @@ def search():
     
     return render_template('admin/search.html', **results)
 
-@app.route('/export/tasks')
-def export_tasks():
-    firm_id = session['firm_id']
-    format_type = request.args.get('format', 'csv')
-    
-    # Get all tasks for the firm
-    tasks = Task.query.outerjoin(Project).filter(
-        db.or_(
-            Project.firm_id == firm_id,
-            db.and_(Task.project_id.is_(None), Task.firm_id == firm_id)
-        )
-    ).all()
-    
-    if format_type == 'csv':
-        import csv
-        import io
-        
-        output = io.StringIO()
-        writer = csv.writer(output)
-        
-        # Write headers
-        writer.writerow([
-            'Task Title', 'Description', 'Status', 'Priority', 
-            'Due Date', 'Estimated Hours', 'Actual Hours',
-            'Assignee', 'Project', 'Client', 'Created Date'
-        ])
-        
-        # Write data
-        for task in tasks:
-            writer.writerow([
-                task.title,
-                task.description or '',
-                task.status,
-                task.priority,
-                task.due_date.strftime('%Y-%m-%d') if task.due_date else '',
-                task.estimated_hours or '',
-                task.actual_hours or '',
-                task.assignee.name if task.assignee else '',
-                task.project.name if task.project else 'Independent Task',
-                task.project.client_name if task.project else '',
-                task.created_at.strftime('%Y-%m-%d %H:%M:%S') if task.created_at else ''
-            ])
-        
-        output.seek(0)
-        
-        from flask import Response
-        return Response(
-            output.getvalue(),
-            mimetype='text/csv',
-            headers={'Content-Disposition': f'attachment; filename=tasks_export_{date.today().strftime("%Y%m%d")}.csv'}
-        )
-    
-    return jsonify({'error': 'Unsupported format'}), 400
+# Export routes moved to export blueprint
 
-@app.route('/export/projects')
-def export_projects():
-    firm_id = session['firm_id']
-    format_type = request.args.get('format', 'csv')
-    
-    # Get all projects for the firm
-    projects = Project.query.filter_by(firm_id=firm_id).all()
-    
-    if format_type == 'csv':
-        import csv
-        import io
-        
-        output = io.StringIO()
-        writer = csv.writer(output)
-        
-        # Write headers
-        writer.writerow([
-            'Project Name', 'Client Name', 'Status', 'Priority',
-            'Start Date', 'Due Date', 'Template', 'Progress %',
-            'Total Tasks', 'Completed Tasks', 'Created Date'
-        ])
-        
-        # Write data
-        for project in projects:
-            completed_tasks = len([t for t in project.tasks if t.status == 'Completed'])
-            total_tasks = len(project.tasks)
-            
-            writer.writerow([
-                project.name,
-                project.client_name,
-                project.status,
-                project.priority if hasattr(project, 'priority') else 'Medium',
-                project.start_date.strftime('%Y-%m-%d') if project.start_date else '',
-                project.due_date.strftime('%Y-%m-%d') if project.due_date else '',
-                project.template_origin.name if project.template_origin else '',
-                f'{project.progress_percentage}%',
-                total_tasks,
-                completed_tasks,
-                project.created_at.strftime('%Y-%m-%d %H:%M:%S') if project.created_at else ''
-            ])
-        
-        output.seek(0)
-        
-        from flask import Response
-        return Response(
-            output.getvalue(),
-            mimetype='text/csv',
-            headers={'Content-Disposition': f'attachment; filename=projects_export_{date.today().strftime("%Y%m%d")}.csv'}
-        )
-    
-    return jsonify({'error': 'Unsupported format'}), 400
-
-@app.route('/export/clients')
-def export_clients():
-    firm_id = session['firm_id']
-    format_type = request.args.get('format', 'csv')
-    
-    # Get all clients for the firm
-    clients = Client.query.filter_by(firm_id=firm_id).all()
-    
-    if format_type == 'csv':
-        import csv
-        import io
-        
-        output = io.StringIO()
-        writer = csv.writer(output)
-        
-        # Write headers
-        writer.writerow([
-            'Client Name', 'Entity Type', 'Email', 'Phone', 'Address',
-            'Contact Person', 'Tax ID', 'Active Projects', 'Notes', 'Created Date'
-        ])
-        
-        # Write data
-        for client in clients:
-            active_projects = Project.query.filter_by(client_id=client.id, status='Active').count()
-            
-            writer.writerow([
-                client.name,
-                client.entity_type or '',
-                client.email or '',
-                client.phone or '',
-                client.address or '',
-                client.contact_person or '',
-                client.tax_id or '',
-                active_projects,
-                client.notes or '',
-                client.created_at.strftime('%Y-%m-%d %H:%M:%S') if client.created_at else ''
-            ])
-        
-        output.seek(0)
-        
-        from flask import Response
-        return Response(
-            output.getvalue(),
-            mimetype='text/csv',
-            headers={'Content-Disposition': f'attachment; filename=clients_export_{date.today().strftime("%Y%m%d")}.csv'}
-        )
-    
-    return jsonify({'error': 'Unsupported format'}), 400
-
-@app.route('/clients/<int:id>/edit', methods=['GET', 'POST'])
-def edit_client(id):
-    client = Client.query.get_or_404(id)
-    if client.firm_id != session['firm_id']:
-        flash('Access denied', 'error')
-        return redirect(url_for('clients'))
-    
-    if request.method == 'POST':
-        client.name = request.form.get('name')
-        client.email = request.form.get('email')
-        client.phone = request.form.get('phone')
-        client.address = request.form.get('address')
-        client.contact_person = request.form.get('contact_person')
-        client.entity_type = request.form.get('entity_type')
-        client.tax_id = request.form.get('tax_id')
-        client.notes = request.form.get('notes')
-        
-        db.session.commit()
-        
-        # Activity log
-        create_activity_log(f'Client "{client.name}" updated', session['user_id'])
-        
-        flash('Client updated successfully!', 'success')
-        return redirect(url_for('view_client', id=client.id))
-    
-    return render_template('clients/edit_client.html', client=client)
-
-@app.route('/clients/<int:id>/delete', methods=['POST'])
-def delete_client(id):
-    client = Client.query.get_or_404(id)
-    
-    # Check access permission
-    if client.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        client_name = client.name
-        
-        # Get counts for confirmation message
-        project_count = Project.query.filter_by(client_id=id).count()
-        task_count = Task.query.join(Project).filter_by(client_id=id).count()
-        contact_count = len(client.contacts)
-        
-        # Create activity log before deletion
-        create_activity_log(f'Client "{client_name}" deleted (with {project_count} projects, {task_count} tasks, {contact_count} contacts)', session['user_id'])
-        
-        # Delete all associated projects (which will cascade delete tasks and attachments)
-        projects = Project.query.filter_by(client_id=id).all()
-        for project in projects:
-            # Delete project attachments
-            Attachment.query.filter_by(project_id=project.id).delete()
-            
-            # Delete all tasks and their comments for this project
-            tasks = Task.query.filter_by(project_id=project.id).all()
-            for task in tasks:
-                TaskComment.query.filter_by(task_id=task.id).delete()
-                db.session.delete(task)
-            
-            # Delete the project
-            db.session.delete(project)
-        
-        # Delete any independent tasks for this client
-        independent_tasks = Task.query.filter_by(client_id=id, project_id=None).all()
-        for task in independent_tasks:
-            TaskComment.query.filter_by(task_id=task.id).delete()
-            db.session.delete(task)
-        
-        # Delete client-contact associations
-        ClientContact.query.filter_by(client_id=id).delete()
-        
-        # Delete client attachments
-        Attachment.query.filter_by(client_id=id).delete()
-        
-        # Delete the client itself
-        db.session.delete(client)
-        db.session.commit()
-        
-        return jsonify({
-            'success': True, 
-            'message': f'Client "{client_name}" and all associated data deleted successfully',
-            'redirect': '/clients'
-        })
-    
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'success': False, 'message': f'Error deleting client: {str(e)}'}), 500
-
-@app.route('/clients/<int:id>/mark_inactive', methods=['POST'])
-def mark_client_inactive(id):
-    client = Client.query.get_or_404(id)
-    
-    # Check access permission
-    if client.firm_id != session['firm_id']:
-        return jsonify({'success': False, 'message': 'Access denied'}), 403
-    
-    try:
-        client_name = client.name
-        previous_status = "Active" if client.is_active else "Inactive"
-        
-        # Toggle active status
-        client.is_active = not client.is_active
-        new_status = "Active" if client.is_active else "Inactive"
-        
-        # Update all associated projects to match client status
-        projects = Project.query.filter_by(client_id=id).all()
-        for project in projects:
-            if client.is_active:
-                # If activating client, set projects to Active if they were inactive
-                if project.status in ['Inactive', 'On Hold']:
-                    project.status = 'Active'
-            else:
-                # If deactivating client, set active projects to On Hold
-                if project.status == 'Active':
-                    project.status = 'On Hold'
-        
-        db.session.commit()
-        
-        # Create activity log
-        action = f'Client "{client_name}" marked as {new_status} (was {previous_status})'
-        if not client.is_active:
-            action += f' - {len(projects)} projects set to On Hold'
-        else:
-            action += f' - {len(projects)} projects reactivated'
-        
-        create_activity_log(action, session['user_id'])
-        
-        return jsonify({
-            'success': True, 
-            'message': f'Client "{client_name}" marked as {new_status}',
-            'new_status': new_status,
-            'is_active': client.is_active
-        })
-    
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'success': False, 'message': f'Error updating client status: {str(e)}'}), 500
+# Additional client routes moved to clients blueprint
 
 @app.route('/admin/work_types', methods=['GET'])
 def admin_work_types():
